fix: ensure proper type conversion in period calculation

+++ chisurf/fluorescence/tcspc/convolve.py
@@ -151,7 +151,7 @@ def convolve_lifetime_spectrum_periodic_nb(
     start = max(start, 0)
 
     n_exp = lifetime_spectrum.shape[0] // 2
-    period_n = ceil(period / dt - 0.5)
+    period_n = int(ceil(float(period) / float(dt) - 0.5))
 
     for i in range(start, stop):
         decay[i] = 0
